fix: moved the api files to a separate directory
