+++ src/layout/FeedsLayout.tsx
@@ -3,7 +3,7 @@ import Header from "@/layout/Header";
 import { Outlet } from "react-router";
 import Sidebar from "@/views/Sidebar/Sidebar";
 import { useAppSelector } from "@/store/hooks";
-import { useWhoAmIQuery } from "@/store/authApi";
+import { useWhoAmIQuery } from "@/store/api/authApi";
 
 function FeedsLayout(): ReactElement {
   const [isSidebarOpen, setIsSidebarOpen] = useState(true);
