+++ src/store/api/authApi.ts
@@ -1,7 +1,7 @@
 import { createApi } from "@reduxjs/toolkit/query/react";
 import { IAccountDTO } from "@/common/interfaces/dto/account/iaccount.interface";
 import { ILoginRequestDTO } from "@/common/interfaces/dto/auth/iadmin-login-request.interface";
-import baseQueryWithReAuth from "@/store/baseQueryWithReAuth";
+import baseQueryWithReAuth from "@/store/api/baseQueryWithReAuth";
 import { ILoginResponseDTO } from "@/common/interfaces/dto/auth/ilogin-response.interfaces";
 
 export const authApi = createApi({
