+++ src/store/store.ts
@@ -1,10 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
 import { setupListeners } from "@reduxjs/toolkit/query";
 import rootReducer from "@/store/reducers";
-import { chatsApi } from "@/store/chatsApi";
-import { authApi } from "@/store/authApi";
-import { messagesApi } from "@/store/messagesApi";
-import { upworkApi } from "@/store/upworkApi";
+import { chatsApi } from "@/store/api/chatsApi";
+import { authApi } from "@/store/api/authApi";
+import { messagesApi } from "@/store/api/messagesApi";
+import { upworkApi } from "@/store/api/upworkApi";
 
 const store = configureStore({
   reducer: rootReducer,
